# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: WhatchUp
  description:  WhatchUp (API para la gestión de reportes de seguridad ciudadana y bienestar comunitario.
  version: 2.0.0
servers:
  - url: 'http://localhost:8080'
    description: Servidor de la API demo
paths:
  /auth/login:
    post:
      summary: Iniciar sesión (Autenticar un usuario).
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "Autenticación exitosa"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: "Credenciales inválidas"
        "500":
          description: "Error en el servidor"
  /auth/register:
    post:
      summary: Registrar un nuevo usuario
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: "Usuario registrado exitosamente"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Datos de registro inválidos"
        "500":
          description: "Error en el servidor"
  /auth/activate:
      post:
        summary: Activar cuenta de usuario
        operationId: activateAccount
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountActivationRequest"
        responses:
          "200":
            description: "Cuenta activada exitosamente"
          "400":
            description: "Token de activación inválido"
          "500":
            description: "Error en el servidor"
  /auth/resend-activation:
      post:
        summary: Reenviar email de activación
        operationId: resendActivation
        parameters:
          - name: email
            in: query
            required: true
            schema:
              type: string
        responses:
          "200":
            description: "Email de activación reenviado exitosamente"
          "404":
            description: "Email no encontrado"
          "500":
            description: "Error en el servidor"
  /auth/forgot-password:
      post:
        summary: Solicitar restablecimiento de contraseña
        operationId: forgotPassword
        parameters:
          - name: email
            in: query
            required: true
            schema:
              type: string
        responses:
          "200":
            description: "Email de restablecimiento enviado exitosamente"
          "404":
            description: "Email no encontrado"
          "500":
            description: "Error en el servidor"
  /auth/reset-password:
      post:
        summary: Restablecer contraseña
        operationId: resetPassword
        parameters:
          - name: token
            in: query
            required: true
            schema:
              type: string
          - name: newPassword
            in: query
            required: true
            schema:
              type: string
        responses:
          "200":
            description: "Contraseña restablecida exitosamente"
          "400":
            description: "Token inválido o expirado"
          "500":
            description: "Error en el servidor"
  /users:
    get:
      summary: Obtener todos los usuarios
      operationId: getUsers
      parameters:
        - name: name
          in: query
          description: Filtrar por nombre
          required: false
          schema:
            type: string
        - name: email
          in: query
          description: Filtrar por email
          required: false
          schema:
            type: string
        - name: registerDate
          in: query
          description: Filtrar por fecha de registro
          required: false
          schema:
            type: string
            format: date
        - name: age
          in: query
          description: Filtrar por edad
          required: false
          schema:
            type: integer
        - name: status
          in: query
          description: Filtrar por estado (active/inactive)
          required: false
          schema:
            type: string
            enum: [ active, deleted, registered ]
        - name: order
          in: query
          description: Orden ascendente (asc) o descendente (desc)
          required: false
          schema:
            type: string
            enum: [ asc, desc ]
            default: asc
        - name: page
          in: query
          description: Número de página (empezando desde 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Cantidad de elementos por página
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de usuarios obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaginatedUserResponse"
        '401':
          description: "No autorizado"
        '403':
          description: "No hay permisos para esta accion"
        '500':
          description: "Error en el servidor"
  /users/{id}:
    get:
      summary: Obtener un usuario por ID
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Usuario encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: "Usuario no encontrado"
    patch:
      summary: Actualizar contraseña de un usuario.
      operationId: updatePassword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdateRequest'
      responses:
        "200":
          description: "Contraseña actualizada con éxito"
        "400":
          description: "Error al actualizar la contraseña"
        "404":
          description: "Usuario no encontrado"
    put:
      summary: Actualizar información de un usuario.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: "Usuario actualizado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Error en los datos enviados"
        "404":
          description: "Usuario no encontrado"
    delete:
      summary: Desactivar (eliminar lógicamente) un usuario.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Usuario desactivado correctamente."
        "403":
          description: "No tienes permisos para desactivar usuarios."
        "404":
          description: "Usuario no encontrado."
...
components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ USER,ADMIN ]
        status:
          type: string
          enum: [ REGISTERED, ACTIVE, INACTIVE ]
    RegisterRequest:
      type: object
      required: ["fullName","city","phone","address","email","password"]
      properties:
        fullName:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginRequest:
      type: object
      required: ["email","password"]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    PaginatedUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        paginated:
          $ref: '#/components/schemas/PaginatedContent'
    PasswordUpdateRequest:
      type: object
      required: ["currentPassword","newPassword"]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
    UserUpdateRequest:
      type: object
      required: []
      properties:
        fullName:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
...
    PaginatedContent:
      type: object
      properties:
        totalPages:
          type: integer
          description: Número total de páginas
        totalElements:
          type: integer
          description: Número total de elementos
        currentPage:
          type: integer
          description: Página actual
        pageSize:
          type: integer
          description: Tamaño de la página
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
    AccountActivationRequest:
      type: object
      required:
        - activationCode
        - email
      properties:
        activationCode:
          type: string
        email:
          type: string
          format: email